requirements.txt
flask==3.0.0
flask-cors==4.0.0
google-generativeai==0.3.2
pdfplumber==0.10.3
docx2txt==0.8
pyngrok==7.0.5
werkzeug==3.0.1
Flask
gunicorn
requests
